Move debrack() to utils.js, remove unused vocab/rdf.js functions
