+++ lib/acl.js
@@ -7,7 +7,6 @@ var url = require('url')
 var async = require('async')
 var debug = require('./debug').ACL
 var utils = require('./utils.js')
-var rdfVocab = require('./vocab/rdf.js')
 
 // TODO should this be set?
 process.env.NODE_TLS_REJECT_UNAUTHORIZED = '0'
@@ -307,7 +306,7 @@ function getUserId (req, callback) {
     return callback(null, req.session.userId)
   }
 
-  var delegator = rdfVocab.debrack(onBehalfOf)
+  var delegator = utils.debrack(onBehalfOf)
   verifyDelegator(req.hostname, delegator, req.session.userId,
     function (err, res) {
       if (err) {
@@ -338,7 +337,7 @@ function verifyDelegator (host, ldp, baseUri, delegator, delegatee, callback) {
 
     for (var delegateeIndex in delegatesStatements) {
       var delegateeValue = delegatesStatements[delegateeIndex]
-      if (rdfVocab.debrack(delegateeValue.toString()) === delegatee) {
+      if (utils.debrack(delegateeValue.toString()) === delegatee) {
         callback(null, true)
       }
     }
