+++ lib/utils.js
@@ -11,12 +11,26 @@ exports.serialize = serialize
 exports.translate = translate
 exports.stringToStream = stringToStream
 exports.reqToPath = reqToPath
+exports.debrack = debrack
 
 var fs = require('fs')
 var path = require('path')
 var S = require('string')
 var $rdf = require('rdflib')
 var from = require('from2')
+
+function debrack (s) {
+  if (s.length < 2) {
+    return s
+  }
+  if (s[0] !== '<') {
+    return s
+  }
+  if (s[s.length - 1] !== '>') {
+    return s
+  }
+  return s.substring(1, s.length - 1)
+}
 
 function uriToFilename (uri, base) {
   uri = decodeURIComponent(uri)
